refactor: move SQL-related message catalogue code from core to sqldb
Move MessageCatalogueData, MessageCatalogueRepository, SqlMessageCatalogueService
and related tests from core module to sqldb module where they belong.

This improves architecture by:
- Keeping core module focused on domain abstractions
- Centralizing SQL-related infrastructure in sqldb module
- Making sqldb the shared foundation for postgresql and mysql implementations
