+++ sqldb-mysql/src/main/scala/works/iterative/sqldb/mysql/migration/MySQLMessageCatalogueMigrationCLI.scala
@@ -5,7 +5,7 @@ package works.iterative.sqldb.mysql.migration
 
 import zio.*
 import works.iterative.core.Language
-import works.iterative.core.repository.MessageCatalogueRepository
+import works.iterative.sqldb.MessageCatalogueRepository
 import works.iterative.sqldb.mysql.*
 
 object MySQLMessageCatalogueMigrationCLI extends ZIOAppDefault:
