+++ sqldb-mysql/src/test/scala/works/iterative/sqldb/mysql/MySQLMessageCatalogueMigrationSpec.scala
@@ -7,7 +7,7 @@ import zio.*
 import zio.test.*
 import zio.test.TestAspect.*
 import works.iterative.core.{Language, MessageId}
-import works.iterative.core.repository.MessageCatalogueRepository
+import works.iterative.sqldb.MessageCatalogueRepository
 import works.iterative.sqldb.mysql.testing.MySQLTestingLayers.*
 import works.iterative.sqldb.FlywayMigrationService
 import works.iterative.sqldb.mysql.migration.MySQLMessageCatalogueMigration
