+++ sqldb-postgresql/src/main/scala/works/iterative/sqldb/postgresql/PostgreSQLMessageCatalogue.scala
@@ -5,7 +5,7 @@ package works.iterative.sqldb.postgresql
 
 import com.augustnagro.magnum.*
 import works.iterative.core.{Language, MessageId}
-import works.iterative.core.model.MessageCatalogueData
+import works.iterative.sqldb.MessageCatalogueData
 import java.time.{Instant, OffsetDateTime, ZoneOffset}
 
 given DbCodec[Instant] =
