+++ sqldb-postgresql/src/main/scala/works/iterative/sqldb/postgresql/migration/PostgreSQLMessageCatalogueMigration.scala
@@ -6,7 +6,7 @@ package works.iterative.sqldb.postgresql.migration
 import zio.*
 import zio.json.*
 import works.iterative.core.{Language, MessageId}
-import works.iterative.core.repository.MessageCatalogueRepository
+import works.iterative.sqldb.MessageCatalogueRepository
 import works.iterative.sqldb.postgresql.MessageCatalogue
 
 object MessageCatalogueMigration:
