+++ sqldb-postgresql/src/test/scala/works/iterative/sqldb/postgresql/MessageCatalogueMigrationSpec.scala
@@ -7,7 +7,7 @@ import zio.*
 import zio.test.*
 import zio.test.TestAspect.*
 import works.iterative.core.{Language, MessageId}
-import works.iterative.core.repository.MessageCatalogueRepository
+import works.iterative.sqldb.MessageCatalogueRepository
 import works.iterative.sqldb.postgresql.testing.PostgreSQLTestingLayers.*
 import works.iterative.sqldb.FlywayMigrationService
 import works.iterative.sqldb.postgresql.migration.MessageCatalogueMigration
