+++ sqldb-postgresql/src/test/scala/works/iterative/sqldb/postgresql/MessageCatalogueRepositorySpec.scala
@@ -8,7 +8,7 @@ import zio.*
 import zio.test.*
 import zio.test.TestAspect.*
 import works.iterative.core.Language
-import works.iterative.core.repository.MessageCatalogueRepository
+import works.iterative.sqldb.MessageCatalogueRepository
 import works.iterative.sqldb.postgresql.testing.PostgreSQLTestingLayers.*
 
 object MessageCatalogueRepositorySpec extends ZIOSpecDefault:
