+++ sqldb-postgresql/src/test/scala/works/iterative/sqldb/postgresql/RealMessageMigrationSpec.scala
@@ -8,8 +8,7 @@ import zio.*
 import zio.test.*
 import zio.test.TestAspect.*
 import works.iterative.core.{Language, MessageId}
-import works.iterative.core.repository.MessageCatalogueRepository
-import works.iterative.core.service.impl.SqlMessageCatalogueService
+import works.iterative.sqldb.{MessageCatalogueRepository, SqlMessageCatalogueService}
 import works.iterative.sqldb.postgresql.testing.PostgreSQLTestingLayers.*
 import works.iterative.sqldb.postgresql.migration.MessageCatalogueMigration
